Revert "Implement consoleRedirect: true"
This reverts commit 64ca79d. It turns out that this wasn't really
necessary after all, as Karma is supposed to redirect this by itself. I
was however somewhat deceived by it not redirecting console.log. This
was a unintentional breaking change of karma-1.5.0 [1].

The workaround for the issue of missing log statements is to override
the default `level` of `browserConsoleLogOptions`, as shown below.

  // karma.conf.js
  module.exports = function(config) {
    config.set({
      browsers: ['jsdom'],

      browserConsoleLogOptions: {
        level: 'log'
      }
    });
  };

[1] https://github.com/karma-runner/karma/issues/2582